fix missing cast error

+++ src/Swastika.Cms.Web.Mvc/Areas/Portal/Controllers/ModuleController.cs
@@ -1,4 +1,8 @@
 // Licensed to the Swastika I/O Foundation under one or more agreements.
+// The Swastika I/O Foundation licenses this file to you under the GNU General Public License v3.0 license.
+// See the LICENSE file in the project root for more information.
+
+// Licensed to the Swastika I/O Foundation under one or more agreements.
 // The Swastika I/O Foundation licenses this file to you under the GNU General Public License v3.0.
 // See the LICENSE file in the project root for more information.
 
@@ -62,7 +66,7 @@ namespace Swastika.Cms.Mvc.Areas.Portal.Controllers
             var Module = new BEModuleViewModel(new SiocModule()
             {
                 Specificulture = CurrentLanguage,
-                Status = Common.Utility.Enums.SWStatus.Published
+                Status = (int)Common.Utility.Enums.SWStatus.Published
             });
             return View(Module);
         }
